ctop/requirements.txt
charset-normalizer==3.4.2
lxml==6.0.0
pillow==11.3.0
python-docx==1.2.0
reportlab==4.4.2
tqdm==4.67.1
typing_extensions==4.14.1
